# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: BridgeYok HTTP API
  version: 0.1.0
  description: Public HTTP contract for the BridgeYok service foundation.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
  /v1/history/boards:
    get:
      operationId: listHistoryBoards
      description: Lists the authenticated user's recent completed boards across casual tables and Team Match rooms. Keyset pagination is ordered by completion time; the response contains only board result and lineup metadata, never a deal.
      security:
      - bearerAuth: []
      parameters:
      - name: cursor
        in: query
        schema:
          type: string
          maxLength: 512
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 40
          default: 16
      responses:
        '200':
          description: Recent authorized board history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryBoardPage'
        default:
          description: Authentication, pagination, or history infrastructure failure.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/matches:
    get:
      operationId: listMatches
      description: Lists at most 20 matches assigned to the authenticated identity,
        unfinished matches first then recent history; never public discovery.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                maxItems: 20
                items:
                  $ref: '#/components/schemas/MatchView'
        default:
          description: Structured authentication, input, membership, state, capacity,
            or infrastructure failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      operationId: createMatch
      description: Creates fresh rooms for eight registered accounts, all unready. The
        owner must be assigned. Identical request IDs return the same match; changed
        bodies conflict. Each assigned account may have at most four pending/active
        matches.
      security:
      - bearerAuth: []
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchView'
        default:
          description: Structured authentication, input, membership, state, capacity,
            or infrastructure failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchCreateRequest'
  /v1/matches/{matchId}:
    get:
      operationId: getMatch
      description: Returns recipient match status and own room. Comparisons are withheld
        until completion.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchView'
        default:
          description: Structured authentication, input, membership, state, capacity,
            or infrastructure failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  schemas:
    MatchSeatRequest:
      type: object
      additionalProperties: false
      required:
      - userId
      - room
      - seat
      properties:
        userId:
          type: string
          format: uuid
        room:
          type: string
          enum:
          - OPEN
          - CLOSED
        seat:
          $ref: '#/components/schemas/TableSeat'
    MatchCreateRequest:
      type: object
      additionalProperties: false
      required:
      - requestId
      - boardCount
      - assignments
      properties:
        requestId:
          type: string
          minLength: 8
          maxLength: 64
          pattern: ^[A-Za-z0-9_-]+$
        boardCount:
          type: integer
          minimum: 1
          maximum: 32
        assignments:
          type: array
          minItems: 8
          maxItems: 8
          items:
            $ref: '#/components/schemas/MatchSeatRequest'
...
    HistoryParticipant:
      type: object
      additionalProperties: false
      required: [id, nickname, isBot]
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
          minLength: 2
          maxLength: 64
        isBot:
          type: boolean
    HistoryPair:
      type: object
      additionalProperties: false
      required: [id, members]
      properties:
        id:
          type: string
        members:
          type: array
          minItems: 2
          maxItems: 2
          items:
            $ref: '#/components/schemas/HistoryParticipant'
    HistoryLineup:
      type: object
      additionalProperties: false
      required: [seats, northSouth, eastWest]
      properties:
        seats:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/HistoryParticipant'
        northSouth:
          $ref: '#/components/schemas/HistoryPair'
        eastWest:
          $ref: '#/components/schemas/HistoryPair'
    HistoryBoard:
      type: object
      additionalProperties: false
      required: [boardId, tableId, boardNumber, result, lineup, viewerSeat, completedAt, replayAllowed]
      properties:
        boardId:
          type: string
          format: uuid
        tableId:
          type: string
          format: uuid
        boardNumber:
          type: integer
          minimum: 1
        result:
          $ref: '#/components/schemas/ReplayResult'
        lineup:
          $ref: '#/components/schemas/HistoryLineup'
        viewerSeat:
          $ref: '#/components/schemas/TableSeat'
        completedAt:
          type: string
          format: date-time
        matchId:
          type: string
          format: uuid
        matchStatus:
          type: string
          enum: [WAITING, ACTIVE, COMPLETE, CANCELLED]
        matchRoom:
          type: string
          enum: [OPEN, CLOSED]
        team:
          type: string
          enum: [A, B]
        teamAIMP:
          type: integer
          minimum: -24
          maximum: 24
        replayAllowed:
          type: boolean
    HistoryBoardPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 40
          items:
            $ref: '#/components/schemas/HistoryBoard'
        nextCursor:
          type: string
    MatchComparison:
      type: object
      additionalProperties: false
      required:
      - boardId
      - openScoreNS
      - closedScoreNS
      - teamAIMP
      properties:
        boardId:
          type: string
          format: uuid
        openScoreNS:
          type: integer
        closedScoreNS:
          type: integer
        teamAIMP:
          type: integer
          minimum: -24
          maximum: 24
    MatchView:
      type: object
      additionalProperties: false
      required:
      - id
      - status
      - revision
      - isOwner
      - tableId
      - tableRevision
      - room
      - seat
      - team
      - boardCount
      - ready
      - readyCount
      - openCompleted
      - closedCompleted
      - canStart
      - canCancel
      - syncPending
      - teamAIMP
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - WAITING
          - ACTIVE
          - COMPLETE
          - CANCELLED
        revision:
          type: integer
          format: int64
          minimum: 0
        isOwner:
          type: boolean
        tableId:
          type: string
          format: uuid
        tableRevision:
          type: integer
          format: int64
          minimum: 0
        room:
          type: string
          enum:
          - OPEN
          - CLOSED
        seat:
          $ref: '#/components/schemas/TableSeat'
        team:
          type: string
          enum:
          - A
          - B
        boardCount:
          type: integer
          minimum: 1
          maximum: 32
        ready:
          type: boolean
        readyCount:
          type: integer
          minimum: 0
          maximum: 8
        openCompleted:
          type: integer
          minimum: 0
          maximum: 32
        closedCompleted:
          type: integer
          minimum: 0
          maximum: 32
        canStart:
          type: boolean
        canCancel:
          type: boolean
        syncPending:
          type: boolean
        comparisons:
          type: array
          maxItems: 32
          items:
            $ref: '#/components/schemas/MatchComparison'
        teamAIMP:
          type: integer
          minimum: -768
          maximum: 768
...
    ReplayContract:
      type: object
      additionalProperties: false
      required:
      - level
      - strain
      - doubling
      - declarer
      properties:
        level:
          type: integer
          minimum: 1
          maximum: 7
        strain:
          enum:
          - C
          - D
          - H
          - S
          - NT
          type: string
        doubling:
          enum:
          - UNDOUBLED
          - DOUBLED
          - REDOUBLED
          type: string
        declarer:
          enum:
          - N
          - E
          - S
          - W
          type: string
    ReplayResult:
      type: object
      additionalProperties: false
      required:
      - rulesetVersion
      - passedOut
      - tricksDeclarer
      - tricksNS
      - tricksEW
      - vulnerability
      - scoreNS
      properties:
        rulesetVersion:
          type: string
          minLength: 1
        passedOut:
          type: boolean
        contract:
          $ref: '#/components/schemas/ReplayContract'
        tricksDeclarer:
          type: integer
          minimum: 0
          maximum: 13
        tricksNS:
          type: integer
          minimum: 0
          maximum: 13
        tricksEW:
          type: integer
          minimum: 0
          maximum: 13
        vulnerability:
          enum:
          - NONE
          - NS
          - EW
          - BOTH
          type: string
        scoreNS:
          type: integer
...
    Problem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
          maxLength: 160
        status:
          type: integer
          minimum: 400
          maximum: 599
        requestId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[A-Za-z0-9_-]+$"
        code:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[A-Z][A-Z0-9_]*$"
        messageKey:
          type: string
          minLength: 1
          maxLength: 128
        retryable:
          type: boolean
...
    TableSeat:
      type: string
      enum:
        - N
        - E
        - S
        - W
...